% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\docType{data}
\encoding{UTF-8}
\name{guide}
\alias{guide}
\title{Guide dataframe for plotting smoothed sortSeq data}
\format{Dataframe with 2 rows and 6 variables:

	\describe{
		\item{order}{Order to plot data in}
		\item{name.rep}{Name of replicating sample}
		\item{name.nonRep}{Name of non-replicating sample}
		\item{raw}{Should raw data be plotted?}
		\item{smooth}{Should smooth data be plotted?}
		\item{color}{Color to plot the profile in}
	}}
\usage{
data(guide)
}
\description{
Guide dataframe for plotting smoothed sortSeq data
}
\examples{
data(guide)
}
\keyword{datasets}
\keyword{replication}
\keyword{sortSeq}
