% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLabels.R
\name{makeLabels}
\alias{makeLabels}
\title{A helper function to create axis ticks and human readable labels.}
\usage{
makeLabels(theMin, theMax, unit = "")
}
\arguments{
\item{theMin}{Minimum value for the scale (double).}

\item{theMax}{Maximum value for the scale (double).}

\item{unit}{Unit to use for the labels (string, optional)}
}
\description{
makeLabels is called by plotGenome() and plotCoverage() functions. It creates a dataframe
containing two columns: 'ticks' and 'labels'. 'Ticks' contains axis ticks coordinates, 'labels'
will contain human readable lables for the ticks (using prefixes and optional units).
}
\examples{
labels <- makeLabels(0,1200000,"b")
}
\keyword{bioinformatics}
\keyword{genomics}
\keyword{plotting}
