% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRatio.R
\name{makeRatio}
\alias{makeRatio}
\title{A function to calculate 'score' ratio between two bed dataframes
makeRatio merges two supplied bed dataframes, calculates ratio of their "score" values
normalises the ratio by the 'score' sums.}
\usage{
makeRatio(bedRep, bedNonRep)
}
\arguments{
\item{bedRep}{Bed dataframe containing read counts from a replicating sample (dataframe).
The bed dataframe must contain "chrom","chromStart", "chromEnd" and "score" columns.}

\item{bedNonRep}{Bed dataframe containing read counts from a non-replicating sample (dataframe).
The bed dataframe must contain "chrom","chromStart", "chromEnd" and "score" columns.}
}
\description{
A function to calculate 'score' ratio between two bed dataframes
makeRatio merges two supplied bed dataframes, calculates ratio of their "score" values
normalises the ratio by the 'score' sums.
}
\examples{
ratioDF <- makeRatio(W303_S,W303_G2)
}
\keyword{bioinformatics}
\keyword{genomics}
\keyword{replication}
