% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatio.R
\name{plotRatio}
\alias{plotRatio}
\title{A function to plot a histogram of supplied ratio vector
plotRatio plots histogram of values in a supplied vector using ggplot2 and
highlights interval between 1 and 2 in green.}
\usage{
plotRatio(ratio, plotting = TRUE)
}
\arguments{
\item{ratio}{A numeric vector containing raw or smoothed ratio values (vector).}

\item{plotting}{Should the plot object be sent to the default device? (boolean, defaults to TRUE).}
}
\description{
A function to plot a histogram of supplied ratio vector
plotRatio plots histogram of values in a supplied vector using ggplot2 and
highlights interval between 1 and 2 in green.
}
\examples{
plotRatio(W303$ratio)
plotObject <- plotRatio(W303$ratio,plotting=FALSE)
}
\keyword{BED}
\keyword{bioinformatics}
\keyword{genomics}
\keyword{histogram}
