% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmChr.R
\name{rmChr}
\alias{rmChr}
\title{A function to remove single chromosome data from a bed dataframe}
\usage{
rmChr(bed, chr)
}
\arguments{
\item{bed}{A bed dataframe containing 'chrom' column (dataframe, required).}

\item{chr}{Chromosome to remove (string, required).}
}
\description{
A function to remove single chromosome data from a bed dataframe
}
\examples{
bedDF <- rmChr(W303_S,"chrM") ## removes mitochondria
}
\keyword{BED}
\keyword{bioinformatics}
\keyword{genomics}
\keyword{outliers}
