% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runGUI}
\alias{runGUI}
\title{A function to launch Repliscope in interactive mode (Shiny app).}
\usage{
runGUI()
}
\description{
A function to launch Repliscope in interactive mode (Shiny app).
}
\keyword{bioinformatics}
\keyword{genomics}
\keyword{replication}
\keyword{shiny}
