% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacCer3.R
\docType{data}
\encoding{UTF-8}
\name{sacCer3}
\alias{sacCer3}
\title{S.cerevisiae genome information}
\format{List containing three dataframes

	\describe{
		\item{genome}{Chromosome information dataframe}
		\item{cen}{Centromere information dataframe}
		\item{ori}{Replication origin information dataframe}
	}}
\source{
Replication origin information: (\href{http://cerevisiae.oridb.org/}{OriDB})
}
\usage{
data(sacCer3)
}
\description{
sacCer3 meta information: chromosome sizes, 
centromere and replication origin positions.
}
\examples{
data(sacCer3)
}
\references{
Siow et al. (2011) NAR 40(Database issue):D682-6
(\href{https://www.ncbi.nlm.nih.gov/pubmed/22121216}{PubMed})
}
\keyword{datasets}
\keyword{replication}
\keyword{sacCer3}
