% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimRatio.R
\name{trimRatio}
\alias{trimRatio}
\title{A function to remove outliers from the "ratio" column of a supplied ratio dataframe
trimRatio is applied to the calculated ratio of read counts from a replicating to
a non-replicating samples.}
\usage{
trimRatio(ratioDF, loLim, hiLim)
}
\arguments{
\item{ratioDF}{A ratio dataframe containing 'ratio' column (dataframe, required).}

\item{loLim}{Low limit threshold (double, required).}

\item{hiLim}{High limit threshold (double, required).}
}
\description{
A function to remove outliers from the "ratio" column of a supplied ratio dataframe
trimRatio is applied to the calculated ratio of read counts from a replicating to
a non-replicating samples.
}
\examples{
W303 <- trimRatio(W303,0.5,1.5)
}
\keyword{bioinformatics}
\keyword{genomics}
\keyword{replication}
